refactor nester directory structure
