+++ src/nester/core/__init__.py
@@ -1,0 +1,1 @@
+"""This module contains all of Nester's core functionality."""
