+++ src/nester/core/templates/c/c_layout.json
