+++ src/nester/core/templates/cpp/cpp_layout.json
